don't pass -file if no PWADs are selected; remove debug line
The old version would pass -file to gzdoom at the end, whether
there were any PWADs selected or not. This didn't seem to matter,
but it was unnecessary and could cause problems later on if
gzdoom's parameter checking gets stricter. Also, I forgot to
remove "echo $pwad" before uploading, so I got rid of it now
to make stdout a bit cleaner.

Signed-off-by: RogueAI42 <declanhoare@exemail.com.au>